+++ backend/db/models/spot.js
@@ -64,7 +64,7 @@ module.exports = (sequelize, DataTypes) => {
       }
     },
     lat: {
-      type: DataTypes.FLOAT(9, 6),
+      type: DataTypes.NUMERIC(9, 6),
       allowNull: false,
       validate: {
         min: -90,
@@ -72,7 +72,7 @@ module.exports = (sequelize, DataTypes) => {
       }
     },
     lng: {
-      type: DataTypes.FLOAT(9, 6),
+      type: DataTypes.NUMERIC(9, 6),
       allowNull: false,
       validate: {
         min: -180,
